Refactor structure of jmh package.
