+++ src/main/java/comparator/jmh/JMHResults.java
@@ -2,6 +2,13 @@ package comparator.jmh;
 
 import comparator.Artifact;
 import comparator.Results;
+import comparator.jmh.metrics.JMHAllocRateNorm;
+import comparator.jmh.metrics.JMHInstructions;
+import comparator.jmh.metrics.JMHMemoryLoads;
+import comparator.jmh.metrics.JMHMemoryStores;
+import comparator.jmh.metrics.JMHPerfResults;
+import comparator.jmh.metrics.JMHPrimaryScore;
+
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.io.PrintWriter;
