+++ src/main/java/comparator/jmh/launch/output/perf/PerfMemoryEvents.java
@@ -1,4 +1,4 @@
-package comparator.jmh.perf;
+package comparator.jmh.launch.output.perf;
 
 import java.io.IOException;
