+++ src/main/java/comparator/jmh/metrics/JMHMetric.java
@@ -1,4 +1,4 @@
-package comparator.jmh;
+package comparator.jmh.metrics;
 
 import comparator.Artifact;
