+++ src/test/java/comparator/comparison/JITResultsComparisonTest.java
@@ -4,12 +4,12 @@ import comparator.Artifact;
 import comparator.JITResults;
 import comparator.jitlog.LogResults;
 import comparator.jitlog.test.JITLogFixture;
-import comparator.jmh.JMHAllocRateNorm;
-import comparator.jmh.JMHInstructions;
-import comparator.jmh.JMHMemoryLoads;
-import comparator.jmh.JMHMemoryStores;
-import comparator.jmh.JMHPrimaryScore;
 import comparator.jmh.JMHResults;
+import comparator.jmh.metrics.JMHAllocRateNorm;
+import comparator.jmh.metrics.JMHInstructions;
+import comparator.jmh.metrics.JMHMemoryLoads;
+import comparator.jmh.metrics.JMHMemoryStores;
+import comparator.jmh.metrics.JMHPrimaryScore;
 import comparator.method.TargetMethod;
 import java.nio.file.Path;
 import java.util.List;
