+++ src/test/java/comparator/comparison/StubResults.java
@@ -3,10 +3,10 @@ package comparator.comparison;
 import comparator.Artifact;
 import comparator.JITResults;
 import comparator.jitlog.LogResults;
-import comparator.jmh.JMHAllocRateNorm;
-import comparator.jmh.JMHPrimaryScore;
 import comparator.jmh.JMHResults;
 import comparator.jmh.fixtures.JMHTarget;
+import comparator.jmh.metrics.JMHAllocRateNorm;
+import comparator.jmh.metrics.JMHPrimaryScore;
 import comparator.method.TargetMethod;
 import java.nio.file.Path;
 import java.util.List;
