+++ src/test/java/comparator/jmh/JMHAllocRateNormTest.java
@@ -3,6 +3,8 @@ package comparator.jmh;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import comparator.jmh.metrics.JMHAllocRateNorm;
+
 class JMHAllocRateNormTest {
     @Test
     void comparesAllocRateUsingAccuracy() {
