+++ src/test/java/comparator/jmh/JMHInstructionsTest.java
@@ -3,6 +3,8 @@ package comparator.jmh;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import comparator.jmh.metrics.JMHInstructions;
+
 class JMHInstructionsTest {
     @Test
     void comparesInstructionsUsingAccuracy() {
